+++ zb_io/src/cellar/mod.rs
@@ -1,5 +1,7 @@
+pub mod bottle_prefix;
 pub mod link;
 pub mod materialize;
 
+pub use bottle_prefix::install_bottle_prefix_files;
 pub use link::{LinkedFile, Linker};
 pub use materialize::{Cellar, CopyStrategy, MaterializedKeg};